// src/shared/api/generated/schemas/publicCategoryResponseDto.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Zemlo API
 * Zemlo backend API documentation
 * OpenAPI spec version: 1.0
 */
import type { PublicCategoryChildResponseDto } from './publicCategoryChildResponseDto';

export interface PublicCategoryResponseDto {
  id: string;
  name: string;
  slug: string;
  /** @nullable */
  description: string | null;
  /** @nullable */
  image: string | null;
  /** @nullable */
  parentId: string | null;
  children: PublicCategoryChildResponseDto[];
}