// src/shared/api/generated/schemas/publicProductDetailResponseDto.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * Zemlo API
 * Zemlo backend API documentation
 * OpenAPI spec version: 1.0
 */
import type { PublicBrandDetailResponseDto } from './publicBrandDetailResponseDto';
import type { PublicCategorySummaryResponseDto } from './publicCategorySummaryResponseDto';
import type { PublicProductImageResponseDto } from './publicProductImageResponseDto';
import type { PublicProductVariantResponseDto } from './publicProductVariantResponseDto';

export interface PublicProductDetailResponseDto {
  id: string;
  name: string;
  slug: string;
  /** @nullable */
  description: string | null;
  /** @nullable */
  shortDescription: string | null;
  /** @nullable */
  sku: string | null;
  price: number;
  /** @nullable */
  compareAtPrice: number | null;
  stock: number;
  trackInventory: boolean;
  allowBackorder: boolean;
  hasVariants: boolean;
  isFeatured: boolean;
  /** @nullable */
  weight: number | null;
  /** @nullable */
  length: number | null;
  /** @nullable */
  width: number | null;
  /** @nullable */
  height: number | null;
  keywords: string[];
  /** @nullable */
  metaTitle: string | null;
  /** @nullable */
  metaDescription: string | null;
  category: PublicCategorySummaryResponseDto;
  /** @nullable */
  brand: PublicBrandDetailResponseDto | null;
  images: PublicProductImageResponseDto[];
  variants: PublicProductVariantResponseDto[];
}