// src/shared/api/generated/schemas/publicProductListItemResponseDto.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * Zemlo API
 * Zemlo backend API documentation
 * OpenAPI spec version: 1.0
 */
import type { PublicBrandSummaryResponseDto } from './publicBrandSummaryResponseDto';
import type { PublicCategorySummaryResponseDto } from './publicCategorySummaryResponseDto';
import type { PublicProductImageResponseDto } from './publicProductImageResponseDto';
import type { PublicProductVariantResponseDto } from './publicProductVariantResponseDto';

export interface PublicProductListItemResponseDto {
  id: string;
  name: string;
  slug: string;
  /** @nullable */
  shortDescription: string | null;
  price: number;
  /** @nullable */
  compareAtPrice: number | null;
  stock: number;
  trackInventory: boolean;
  allowBackorder: boolean;
  hasVariants: boolean;
  isFeatured: boolean;
  keywords: string[];
  /** @nullable */
  metaTitle: string | null;
  /** @nullable */
  metaDescription: string | null;
  category: PublicCategorySummaryResponseDto;
  /** @nullable */
  brand: PublicBrandSummaryResponseDto | null;
  images: PublicProductImageResponseDto[];
  variants: PublicProductVariantResponseDto[];
}